// frontend/src/lib/api/models/Note.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CategoryEnum } from './CategoryEnum';
export type Note = {
    readonly id: number;
    title?: string;
    content?: string;
    category?: CategoryEnum;
    readonly owner_id: number;
    readonly created_at: string;
    readonly updated_at: string;
};