// frontend/src/lib/api/models/CategoryEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * * `random-thoughts` - Random Thoughts
 * * `school` - School
 * * `personal` - Personal
 */
export enum CategoryEnum {
    RANDOM_THOUGHTS = 'random-thoughts',
    SCHOOL = 'school',
    PERSONAL = 'personal',
}