// frontend/src/lib/api/models/PatchedNote.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CategoryEnum } from './CategoryEnum';
export type PatchedNote = {
    readonly id?: number;
    title?: string;
    content?: string;
    category?: CategoryEnum;
    readonly owner_id?: number;
    readonly created_at?: string;
    readonly updated_at?: string;
};